Add Qwen ASR transcription mode
